+++ quiltctl/command/ps.go
@@ -118,14 +118,9 @@ func writeMachines(fd io.Writer, machines []db.Machine) {
 	fmt.Fprintln(w, "MACHINE\tROLE\tPROVIDER\tREGION\tSIZE\tPUBLIC IP\tSTATUS")
 
 	for _, m := range db.SortMachines(machines) {
-		status := "disconnected"
-		if m.Connected {
-			status = "connected"
-		}
-
 		fmt.Fprintf(w, "%v\t%v\t%v\t%v\t%v\t%v\t%v\n",
 			util.ShortUUID(m.StitchID), m.Role, m.Provider, m.Region, m.Size,
-			m.PublicIP, status)
+			m.PublicIP, m.Status)
 	}
 }